feat: 로또 컨트롤러, 입출력 뷰 구현
