+++ src/test/kotlin/lotto/domain/WinningLottoTest.kt
@@ -1,5 +1,6 @@
-package lotto
+package lotto.domain
 
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
@@ -13,14 +14,23 @@ class WinningLottoTest {
         winningLotto.lottoNumbers.size shouldBe 6
     }
 
+    @Test
+    fun `당첨 티켓의 번호는 6개가 아니면 에러가 발생한다`() {
+        shouldThrow<IllegalArgumentException> {
+            WinningLotto(listOf(1, 2, 3, 4))
+        }.also {
+            it.message shouldBe "당첨 번호 입력이 유효하지 않습니다"
+        }
+    }
+
     @ParameterizedTest
     @MethodSource("generateWiningLotto")
     fun `로또 티켓이 당첨 티켓과 몇개가 일치하는지 확인 할 수 있다`(
         lottoTicket: LottoTicket,
         winningLotto: WinningLotto,
-        matchCount: Int,
+        count: Int,
     ) {
-        val count = winningLotto.matchedCount(lottoTicket)
+        val matchCount = winningLotto.calculateMatchCount(lottoTicket)
         matchCount shouldBe count
     }
